feat: move inner classes to upper level
