+++ src/test/java/com/kjh/wms/product/feature/Category.java
@@ -1,0 +1,11 @@
+package com.kjh.wms.product.feature;
+
+public enum Category {
+    ELECTRONICS("전자 제품");
+
+    private final String description;
+
+    Category(final String description) {
+        this.description = description;
+    }
+}
